requirements.txt
requests
beautifulsoup4
flask
python-dotenv
gunicorn
urllib3
